+++ ts/packages/knowPro/src/common.ts
@@ -14,9 +14,6 @@ import { DateTimeRange } from "./dateTimeSchema.js";
 import {
     DateRange,
     IConversation,
-    KnowledgeType,
-    ScoredSemanticRefOrdinal,
-    SemanticRef,
     TextLocation,
     TextRange,
 } from "./interfaces.js";
@@ -110,30 +107,6 @@ export function isInDateRange(outerRange: DateRange, date: Date): boolean {
 
 export function isSearchTermWildcard(searchTerm: SearchTerm): boolean {
     return searchTerm.term.text === "*";
-}
-
-export function facetValueToString(facet: kpLib.Facet): string {
-    const value = facet.value;
-    if (typeof value === "object") {
-        return `${value.amount} ${value.units}`;
-    }
-    return value.toString();
-}
-
-export function* getScoredSemanticRefsFromOrdinals(
-    semanticRefs: SemanticRef[],
-    semanticRefMatches: ScoredSemanticRefOrdinal[],
-    knowledgeType: KnowledgeType,
-): IterableIterator<Scored<SemanticRef>> {
-    for (let semanticRefMatch of semanticRefMatches) {
-        const semanticRef = semanticRefs[semanticRefMatch.semanticRefOrdinal];
-        if (semanticRef.knowledgeType === knowledgeType) {
-            yield {
-                score: semanticRefMatch.score,
-                item: semanticRef,
-            };
-        }
-    }
 }
 
 export async function createConversationFromData(
