+++ ts/packages/knowPro/src/conversationIndex.ts
@@ -21,7 +21,7 @@ import { IndexingEventHandlers } from "./interfaces.js";
 import { conversation as kpLib } from "knowledge-processor";
 import { openai } from "aiclient";
 import { extractKnowledgeFromTextBatch } from "./knowledge.js";
-import { facetValueToString } from "./common.js";
+import { facetValueToString } from "./knowledgeLib.js";
 import { createKnowledgeExtractor } from "./knowledge.js";
 import {
     addToSecondaryIndexes,
