+++ ts/packages/knowPro/src/propertyIndex.ts
@@ -12,7 +12,7 @@ import {
 import { conversation as kpLib } from "knowledge-processor";
 import { IPropertyToSemanticRefIndex } from "./interfaces.js";
 import { TextRangesInScope } from "./collections.js";
-import { facetValueToString } from "./common.js";
+import { facetValueToString } from "./knowledgeLib.js";
 
 export enum PropertyNames {
     EntityName = "name",
